fix: api::Double is double, not long double
mx::api modeled xs:decimal/float/double values as long double, but the
impl only widens core doubles into them, so they never hold extra
precision -- and writing them back via brace-init (core::Decimal{ v })
is a long double -> double narrowing that Clang rejects as ill-formed
while GCC merely warns. That split broke every AppleClang build (the
macOS CMake job and the new Swift package).

Make api::Double a double and sweep the stray literal long double
declarations and their L-suffixed literals to match. xs:* values are
<= 64-bit and long double == double on Apple silicon, so this loses
nothing and removes the narrowing at the root.

+++ src/include/mx/api/ApiCommon.h
@@ -12,7 +12,7 @@ namespace mx
 {
 namespace api
 {
-using Double = long double;
+using Double = double;
 using OptionalDouble = std::optional<Double>;
 constexpr const Double MX_API_EQUALITY_EPSILON = 0.00000001;
